Fixed removeDup function
